refactor: Split select/mod.rs into executor submodules
- Created SelectExecutor impl split across 6 modules:
  * builder.rs - Struct definition and constructors (33 lines)
  * execute.rs - Main execution entry points (103 lines)
  * columns.rs - Column name derivation (113 lines)
  * aggregation.rs - Aggregation and GROUP BY execution (174 lines)
  * nonagg.rs - Non-aggregation execution path (113 lines)
  * utils.rs - Expression analysis utilities (104 lines)
- Converted select/mod.rs to imports and re-exports (39 lines)
- All select tests passing (39 tests)
- No functional changes, pure refactoring

Reduces select/mod.rs from 663 lines to ~40 lines + 6 focused modules
