+++ crates/executor/src/select/executor/aggregation.rs
@@ -1,0 +1,242 @@
+//! Aggregation execution methods for SelectExecutor
+
+use super::builder::SelectExecutor;
+use crate::errors::ExecutorError;
+use crate::evaluator::ExpressionEvaluator;
+use crate::select::cte::CteResult;
+use crate::select::filter::apply_where_filter_basic;
+use crate::select::grouping::{group_rows, AggregateAccumulator};
+use crate::select::helpers::{apply_distinct, apply_limit_offset};
+use std::collections::HashMap;
+
+impl<'a> SelectExecutor<'a> {
+    /// Check if SELECT list contains aggregate functions
+    pub(super) fn has_aggregates(&self, select_list: &[ast::SelectItem]) -> bool {
+        select_list.iter().any(|item| match item {
+            ast::SelectItem::Expression { expr, .. } => self.expression_has_aggregate(expr),
+            _ => false,
+        })
+    }
+
+    /// Check if an expression contains aggregate functions
+    #[allow(clippy::only_used_in_recursion)]
+    pub(super) fn expression_has_aggregate(&self, expr: &ast::Expression) -> bool {
+        match expr {
+            ast::Expression::Function { name, .. } => {
+                matches!(name.to_uppercase().as_str(), "COUNT" | "SUM" | "AVG" | "MIN" | "MAX")
+            }
+            ast::Expression::BinaryOp { left, right, .. } => {
+                self.expression_has_aggregate(left) || self.expression_has_aggregate(right)
+            }
+            _ => false,
+        }
+    }
+
+    /// Execute SELECT with aggregation/GROUP BY
+    pub(super) fn execute_with_aggregation(
+        &self,
+        stmt: &ast::SelectStmt,
+        cte_results: &HashMap<String, CteResult>,
+    ) -> Result<Vec<storage::Row>, ExecutorError> {
+        // Get table
+        let table_name = match &stmt.from {
+            Some(ast::FromClause::Table { name, alias: _ }) => name,
+            Some(_) => {
+                return Err(ExecutorError::UnsupportedFeature(
+                    "Aggregates only supported on single tables".to_string(),
+                ))
+            }
+            None => {
+                return Err(ExecutorError::UnsupportedFeature(
+                    "SELECT without FROM not yet implemented".to_string(),
+                ))
+            }
+        };
+
+        // Check if table is a CTE first, then check database
+        let (schema, rows) = if let Some((cte_schema, cte_rows)) = cte_results.get(table_name) {
+            // Use CTE result
+            (cte_schema.clone(), cte_rows.clone())
+        } else {
+            // Use database table
+            let table = self
+                .database
+                .get_table(table_name)
+                .ok_or_else(|| ExecutorError::TableNotFound(table_name.clone()))?;
+            (table.schema.clone(), table.scan().to_vec())
+        };
+
+        let evaluator = match (self.outer_row, self.outer_schema) {
+            (Some(outer_row), Some(outer_schema)) => {
+                ExpressionEvaluator::with_database_and_outer_context(
+                    &schema,
+                    self.database,
+                    outer_row,
+                    outer_schema,
+                )
+            }
+            _ => ExpressionEvaluator::with_database(&schema, self.database),
+        };
+
+        // Scan and filter rows with WHERE clause
+        let filtered_rows = apply_where_filter_basic(rows, stmt.where_clause.as_ref(), &evaluator)?;
+
+        // Group rows
+        let groups = if let Some(group_by_exprs) = &stmt.group_by {
+            group_rows(&filtered_rows, group_by_exprs, &evaluator)?
+        } else {
+            // No GROUP BY - treat all rows as one group
+            vec![(Vec::new(), filtered_rows)]
+        };
+
+        // Compute aggregates for each group and apply HAVING
+        let mut result_rows = Vec::new();
+        for (group_key, group_rows) in groups {
+            // Compute aggregates for this group
+            let mut aggregate_results = Vec::new();
+            for item in &stmt.select_list {
+                match item {
+                    ast::SelectItem::Expression { expr, .. } => {
+                        let value = self.evaluate_with_aggregates(
+                            expr,
+                            &group_rows,
+                            &group_key,
+                            &evaluator,
+                        )?;
+                        aggregate_results.push(value);
+                    }
+                    ast::SelectItem::Wildcard => {
+                        return Err(ExecutorError::UnsupportedFeature(
+                            "SELECT * not supported with aggregates".to_string(),
+                        ))
+                    }
+                }
+            }
+
+            // Apply HAVING filter
+            let include_group = if let Some(having_expr) = &stmt.having {
+                let having_result = self.evaluate_with_aggregates(
+                    having_expr,
+                    &group_rows,
+                    &group_key,
+                    &evaluator,
+                )?;
+                match having_result {
+                    types::SqlValue::Boolean(true) => true,
+                    types::SqlValue::Boolean(false) | types::SqlValue::Null => false,
+                    other => {
+                        return Err(ExecutorError::InvalidWhereClause(format!(
+                            "HAVING must evaluate to boolean, got: {:?}",
+                            other
+                        )))
+                    }
+                }
+            } else {
+                true
+            };
+
+            if include_group {
+                result_rows.push(storage::Row::new(aggregate_results));
+            }
+        }
+
+        // Apply ORDER BY if present
+        if stmt.order_by.is_some() {
+            return Err(ExecutorError::UnsupportedFeature(
+                "ORDER BY with aggregates not yet implemented".to_string(),
+            ));
+        }
+
+        // Apply DISTINCT if specified
+        let result_rows = if stmt.distinct { apply_distinct(result_rows) } else { result_rows };
+
+        // Don't apply LIMIT/OFFSET if we have a set operation - it will be applied later
+        if stmt.set_operation.is_some() {
+            Ok(result_rows)
+        } else {
+            Ok(apply_limit_offset(result_rows, stmt.limit, stmt.offset))
+        }
+    }
+
+    /// Evaluate an expression in the context of aggregation
+    #[allow(clippy::only_used_in_recursion)]
+    pub(super) fn evaluate_with_aggregates(
+        &self,
+        expr: &ast::Expression,
+        group_rows: &[storage::Row],
+        _group_key: &[types::SqlValue],
+        evaluator: &ExpressionEvaluator,
+    ) -> Result<types::SqlValue, ExecutorError> {
+        match expr {
+            // Aggregate function
+            ast::Expression::Function { name, args } => {
+                let mut acc = AggregateAccumulator::new(name)?;
+
+                // Special handling for COUNT(*)
+                if name.to_uppercase() == "COUNT" && args.len() == 1 {
+                    // Parser represents COUNT(*) as either Wildcard or ColumnRef { column: "*" }
+                    let is_count_star = matches!(args[0], ast::Expression::Wildcard)
+                        || matches!(
+                            &args[0],
+                            ast::Expression::ColumnRef { table: None, column } if column == "*"
+                        );
+
+                    if is_count_star {
+                        // COUNT(*) - count all rows
+                        for _ in group_rows {
+                            acc.accumulate(&types::SqlValue::Integer(1));
+                        }
+                        return Ok(acc.finalize());
+                    }
+                }
+
+                // Regular aggregate - evaluate argument for each row
+                if args.len() != 1 {
+                    return Err(ExecutorError::UnsupportedExpression(format!(
+                        "Aggregate functions expect 1 argument, got {}",
+                        args.len()
+                    )));
+                }
+
+                for row in group_rows {
+                    let value = evaluator.eval(&args[0], row)?;
+                    acc.accumulate(&value);
+                }
+
+                Ok(acc.finalize())
+            }
+
+            // Column reference - use value from group key
+            ast::Expression::ColumnRef { .. } => {
+                // For GROUP BY columns, use the first row's value (they're all the same)
+                if let Some(first_row) = group_rows.first() {
+                    evaluator.eval(expr, first_row)
+                } else {
+                    Ok(types::SqlValue::Null)
+                }
+            }
+
+            // Binary operation
+            ast::Expression::BinaryOp { left, op, right } => {
+                let left_val =
+                    self.evaluate_with_aggregates(left, group_rows, _group_key, evaluator)?;
+                let right_val =
+                    self.evaluate_with_aggregates(right, group_rows, _group_key, evaluator)?;
+
+                // Reuse the binary op evaluation logic from ExpressionEvaluator
+                // Create a temporary evaluator for this
+                let temp_schema = catalog::TableSchema::new("temp".to_string(), vec![]);
+                let temp_evaluator = ExpressionEvaluator::new(&temp_schema);
+                temp_evaluator.eval_binary_op(&left_val, op, &right_val)
+            }
+
+            // Literal
+            ast::Expression::Literal(val) => Ok(val.clone()),
+
+            _ => Err(ExecutorError::UnsupportedExpression(format!(
+                "Unsupported expression in aggregate context: {:?}",
+                expr
+            ))),
+        }
+    }
+}
